fix(runtime): make submitters dyn-safe and move submit config
